improve (WebGPU, MToon): Improve code around parametric rim
We should have the parametric rim result to a temporary variable
